--- components/experimental/spaces/src/interfaces/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export * from "./componentInternalRegistry";
-export * from "./componentRegistryTemplates";
